fix(parse_file): ignore files that aren't UTF-8
I hope this is what this does. Seems like read_to_string should only be
able to succeeded on files that can be read to a utf-8 string, which is
what Rust supports. So anything else (e.g. binary files), will throw
InvalidData and we can just ignore them. Not going to be adding TODOs to
your MIDI files... :)